Changes to Command
Changed the command location and moved it into the Command.class. The PREFIX and INFIX are now static across all commands, but the command and commandDescription are not.

Also made a todo:
In the dispatchers, there needs to be a way to accept no param commands and/or differentiate between them. Example:
">hello" and ">hello F4nt0M"